Move getSeverApp
