+++ modules/firebase/serverApp.ts
@@ -5,25 +5,7 @@
 //
 // SPDX-License-Identifier: MIT
 //
-import { initializeServerApp } from 'firebase/app'
-import { getAuth } from 'firebase/auth'
-import { headers } from 'next/headers'
+import { getServerApp as getServerAppBase } from '@stanfordbdhg/design-system/modules/auth/serverApp'
 import { firebaseConfig } from './config'
 
-export const getServerApp = async () => {
-  const idToken = headers().get('Authorization')?.split('Bearer ').at(1)
-
-  const firebaseServerApp = initializeServerApp(
-    firebaseConfig,
-    idToken ?
-      {
-        authIdToken: idToken,
-      }
-    : {},
-  )
-
-  const auth = getAuth(firebaseServerApp)
-  await auth.authStateReady()
-
-  return { firebaseServerApp, currentUser: auth.currentUser }
-}
+export const getServerApp = () => getServerAppBase(firebaseConfig)
