+++ packages/design-system/package.json
@@ -44,6 +44,7 @@
     "./components/Separator": "./dist/components/Separator/index.js",
     "./molecules/AsideBrandLayout": "./dist/molecules/AsideBrandLayout/index.js",
     "./modules/auth/SignInForm": "./dist//modules/auth/SignInForm/index.js",
+    "./modules/auth/serverApp": "./dist//modules/auth/serverApp.js",
     "./forms/Field": "./dist/forms/Field/index.js",
     "./forms/useForm": "./dist/forms/useForm/index.js",
     "./main.css": "./dist/main.css"
